requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
motor==3.5.1
bcrypt==4.2.0
PyJWT==2.9.0
pydantic==2.9.2
groq==0.9.0
PyPDF2==3.0.1
python-docx==1.1.2
requests==2.32.3
beautifulsoup4==4.12.3
tiktoken==0.7.0
reportlab==4.2.5
tavily-python==0.5.0